Nested for loop integrated
Whilst not directly a 2d array, this nesting of for loops will be handly
for when i integrate more 2d array related work into this. Once again, i
re-interate that this is an application for me to learn a few things, so
it will be a mish mash of various little items, all in one place.