02/09/15
Creation of cell object class. Created a method under main user
interaction, which creates 2d array of objects for the cell class, along
with a testChar variable, to display to prove they're being set to DOA,
upon initialization. @ = dead. # = Alive. Now a second method needs to
be called to update?