Game of Life Complete?
Game of life has been implemented and very close to completion. Rules
governing replication needs re-confirming and examining and also instead
of auto-running, you need to press enter and i've put a delay on it.
This could be to do with limitations of the console. I could look into
taking this off, but it would change too fast.