Nailed it
Conways game of life in a 50x50 2d array. To change array size,
Program.cs and OOP.cs variables need changing. the variables will wither
be currently down as 49 (in OOP.cs) or 50 (Program.cs). An option, for
if the user enters 3, has been added to allow the app to refresh itself
permanently